backend/DermaAI/lambdas/classification/requirements.txt
pillow==11.0.0
torch==2.6.0+cpu
torchvision==0.21.0+cpu
boto3
